pseq: Create dump on runtime pgood failure
Currently no dump is created when a power good failure occurs at
runtime. Add the call to create a dump.

+++ phosphor-power-sequencer/src/ucd90320_monitor.hpp
@@ -122,7 +122,7 @@ class UCD90320Monitor : public PowerSequencerMonitor
      * Reads the mfr_status register
      * @return the register contents
      */
-    uint32_t readMFRStatus();
+    uint64_t readMFRStatus();
 
     /**
      * Reads the status_word register
